OrderItem Integer/Double 改為 BigDecimal 導入精度計算類別
Service 實作「高精度計算」與「資源保障」 (OrderServiceImpl)

計算總額時使用 BigDecimal，並用 StringBuilder 生成發票資訊，最後透過 finally 確保日誌完整。

prodcut 價格欄位改成 BigDecimal

order 總金額欄位改成 BigDecimal

+++ sp2-springboot/src/main/java/com/example/demo/entity/OrderItem.java
@@ -1,5 +1,7 @@
 package com.example.demo.entity;
 
+import java.math.BigDecimal; // 【重點】導入精度計算類別
+
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -22,9 +24,9 @@ public class OrderItem {
 	@ManyToOne
 	@JoinColumn(name = "product_id")
 	private Product product; // 買了什麼商品
-	
+	// 將價格從 Integer/Double 改為 BigDecimal
 	private Integer quantity; 	// 買了多少
-	private Integer price;		// 買的單價
+	private BigDecimal price;		// 買的單價
 	public Integer getId() {
 		return id;
 	}
@@ -49,11 +51,12 @@ public class OrderItem {
 	public void setQuantity(Integer quantity) {
 		this.quantity = quantity;
 	}
-	public Integer getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
-	public void setPrice(Integer price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
+	
 	
 }
